+++ src/euler/sequence/AbundantNumbers.java
@@ -28,7 +28,7 @@ public class AbundantNumbers extends AbstractSequence {
 	}
 
 	@Override
-	public int position() {
+	public long position() {
 		return pos;
 	}
 }
