+++ src/euler/sequence/SquareDigits.java
@@ -53,7 +53,7 @@ public class SquareDigits extends AbstractSequence {
 	}
 
 	@Override
-	public int position() {
+	public long position() {
 		return pos;
 	}
